fix: remove seed data and update ledger sum casting in tests
Removed extensive user, market, outcome, and position seed data from supabase/seed.sql to clean up the initial database setup. Also updated a test in place_bet_test.sql by removing explicit bigint casting on a sum query to improve type handling consistency.

+++ supabase/tests/database/place_bet_test.sql
@@ -86,7 +86,7 @@ set local request.jwt.claim.sub =
 
 select is(
     (
-        select coalesce(sum(delta), 0)::bigint
+        select coalesce(sum(delta), 0)
         from public.ledger
         where profile_id = '90000000-0000-0000-0000-000000000099'
     ),
@@ -108,7 +108,7 @@ select throws_ok(
 
 select is(
     (
-        select coalesce(sum(delta), 0)::bigint
+        select coalesce(sum(delta), 0)
         from public.ledger
         where profile_id = '90000000-0000-0000-0000-000000000099'
     ),
@@ -118,7 +118,7 @@ select is(
 
 select is(
     (
-        select coalesce(sum(delta), 0)::bigint
+        select coalesce(sum(delta), 0)
         from public.ledger
         where profile_id = '90000000-0000-0000-0000-000000000099'
     ),
@@ -140,7 +140,7 @@ select throws_ok(
 
 select is(
     (
-        select coalesce(sum(delta), 0)::bigint
+        select coalesce(sum(delta), 0)
         from public.ledger
         where profile_id = '90000000-0000-0000-0000-000000000099'
     ),
@@ -150,7 +150,7 @@ select is(
 
 select is(
     (
-        select coalesce(sum(delta), 0)::bigint
+        select coalesce(sum(delta), 0)
         from public.ledger
         where profile_id = '90000000-0000-0000-0000-000000000099'
     ),
@@ -172,7 +172,7 @@ select throws_ok(
 
 select is(
     (
-        select coalesce(sum(delta), 0)::bigint
+        select coalesce(sum(delta), 0)
         from public.ledger
         where profile_id = '90000000-0000-0000-0000-000000000099'
     ),
@@ -182,7 +182,7 @@ select is(
 
 select is(
     (
-        select coalesce(sum(delta), 0)::bigint
+        select coalesce(sum(delta), 0)
         from public.ledger
         where profile_id = '90000000-0000-0000-0000-000000000099'
     ),
@@ -204,7 +204,7 @@ select throws_ok(
 
 select is(
     (
-        select coalesce(sum(delta), 0)::bigint
+        select coalesce(sum(delta), 0)
         from public.ledger
         where profile_id = '90000000-0000-0000-0000-000000000099'
     ),
@@ -224,7 +224,7 @@ select lives_ok(
 
 select is(
     (
-        select coalesce(sum(delta), 0)::bigint
+        select coalesce(sum(delta), 0)
         from public.ledger
         where profile_id = '90000000-0000-0000-0000-000000000099'
     ),
@@ -246,7 +246,7 @@ select throws_ok(
 
 select is(
     (
-        select coalesce(sum(delta), 0)::bigint
+        select coalesce(sum(delta), 0)
         from public.ledger
         where profile_id = '90000000-0000-0000-0000-000000000099'
     ),
